Added home link, cleaned up, linted and tested previous commit.
