+++ services/frontend/src/scenes/Profile/ProfileLayout.js
@@ -1,17 +1,27 @@
 import React from "react";
 import {Logo} from "assets/icons";
 import "./Profile.scss";
-import { Button, Tabs, Tab, Typography, AppBar } from '@material-ui/core';
-import { makeStyles } from '@material-ui/styles';
 
-const useStyles = makeStyles({
-  root: {
-    backgroundColor: 'red',
-  },
-});
+const ProfileLayout = () => {
+    return (
+        <div className="profile">
+            <div className="profile-header">
+                <a href="/">
+                    <Logo className="logo-icon" />
+                </a>
 
-export default function ProfileLayoutj() {
-  const classes = useStyles();
-  return <div className={classes.root} />;
-}
+                <div className="profile-controls">
+                    <button className="profile-search">ff</button>
+                    <button className="profile-tools">SIGN UP</button>
+                </div>
+            </div>
 
+            <div className="profile-jumbotron">
+                <h1 className="profile-tagline">WELCOME TO SKILLHUB!</h1>
+                <button className="profile-create-profile">CREATE YOUR PROFILE</button>
+            </div>
+        </div>
+    );
+};
+
+export default ProfileLayout;
